Add runtime error checking to platform, promise based extension.tab methods
