+++ app/scripts/lib/util.js
@@ -1,3 +1,4 @@
+const extension = require('extensionizer')
 const ethUtil = require('ethereumjs-util')
 const assert = require('assert')
 const BN = require('bn.js')
@@ -156,6 +157,24 @@ function mapObjectValues (object, cb) {
   return mappedObject
 }
 
+/**
+ * Returns an Error if extension.runtime.lastError is present
+ * this is a workaround for the non-standard error object thats used
+ * @returns {Error}
+ */
+function checkForError () {
+  const lastError = extension.runtime.lastError
+  if (!lastError) {
+    return
+  }
+  // if it quacks like an Error, its an Error
+  if (lastError.stack && lastError.message) {
+    return lastError
+  }
+  // repair incomplete error object (eg chromium v77)
+  return new Error(lastError.message)
+}
+
 module.exports = {
   removeListeners,
   applyListeners,
@@ -168,4 +187,5 @@ module.exports = {
   BnMultiplyByFraction,
   getRandomArrayItem,
   mapObjectValues,
+  checkForError,
 }
